Fix warnings

+++ src/Edit/Interface/edit_interface.cpp
@@ -568,10 +568,10 @@ edit_interface_rep::apply_changes () {
   // cout << "Handling extents\n";
   if (env_change & (THE_TREE+THE_ENVIRONMENT+THE_EXTENTS)) {
     string medium= get_init_string (PAGE_MEDIUM);
-    SI ex1= (SI) eb->x1*magf;
-    SI ey1= (SI) eb->y1*magf;
-    SI ex2= (SI) eb->x2*magf;
-    SI ey2= (SI) eb->y2*magf;
+    SI ex1= (SI) ((double) eb->x1) * magf;
+    SI ey1= (SI) ((double) eb->y1) * magf;
+    SI ex2= (SI) ((double) eb->x2) * magf;
+    SI ey2= (SI) ((double) eb->y2) * magf;
     abs_round (ex1, ey1);
     abs_round (ex2, ey2);
     SI w, h;
